refactor(napi,wasm): split monolithic binding lib.rs into focused modules (#210)
Both binding crates kept their entire public surface in a single lib.rs
(napi 2175 lines, wasm 1714 lines), the two highest-churn files in the
workspace. Every new client API landed there, concentrating merge
conflicts and review noise.

Mirror the Python crate's layout by splitting each into focused modules:

  lib.rs     module registration + re-exports only
  config.rs  Config / WasmClientConfig
  error.rs   PubMedError -> napi/js error conversion
  models.rs  DTO structs + From impls
  query.rs   SearchQuery wrapper (+ helpers)
  client.rs  unified client methods

napi/wasm both have a single unified client struct (unlike Python's split
PubMedClient/PmcClient), so client methods stay in one client.rs to keep
the generated bindings' method order stable.

Generated output is unchanged: napi index.d.ts is byte-identical and the
wasm-bindgen .d.ts diffs clean against a baseline build. napi vitest suite
(57 tests) passes.
